tugas2/requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Neo4j Driver
neo4j==5.14.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
colorama==0.4.6
